NMS-9225: Added JMS heartbeat outage test.
